fix: update data/dataset.py module for new CSV format
With the change from train0 to train1, I have changed the CSV format to include the full class name (e.g. car make/model/year) instead of just the class ID. Doing so changed the field called "label" to "class_id", so changed had to be made in the data/dataset.py module to parse through the dataset CSV properly.